[FEAT] 사용자 정보 페이지 네트워크 연결 (#55)
* gradle: splash 의존성 추가

* feat: splash 화면 연결

* refactor: 스플래시 api 의존성 삭제

* gradle: minSDK 버전 수정

* feat: 스플래시 화면 연결

* chore: ktFormat

* feat: 유저 정보를 불러오는 api 서비스 추가

* feat: 유저 정보 dataSource, repository 생성

* feat: 유저 정보와 UI 연결

* feat: 유저 거래정보 UI 연결

* refactor: 유저 페이지 stringRes로 관리

* refactor: 유저 response 패키지 이동

* refactor: 유저 source,repository 분리

* refactor: ui에 필요한 정보만 넘기도록 수정

* chore: ktFormat

* refactor: Profile prefix 네이밍 변경

* fix : 뒤로가기 시, 스플래시 화면 보이는 문제 해결

* chore: ktFormat

+++ app/src/main/java/com/example/swapit/ui/user/UserProfileItem.kt
@@ -20,7 +20,7 @@ import com.example.swapit.ui.theme.Typography
 @Composable
 fun ProfileItem(
     text: String,
-    count: Int,
+    count: Long,
 ) {
     Row(
         modifier =
